Move PagerTransformer to Random package
